backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.1.3

black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0

fastapi==0.110.1
flake8==7.3.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6

mypy==1.19.1
numpy==2.4.3
pandas==3.0.1
passlib==1.7.4
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22

requests==2.32.5
rich==14.3.3
sniffio==1.3.1
starlette==0.37.2

supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3

typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0